Move the 37 tools into a tools/ package by domain
Step 4, the bulk of the split. server.py goes from 2327 lines to 162:
seven domain modules under tools/, each decorating against the shared mcp
object so every tool name is unchanged.

The tool functions and DEFAULT_SETTINGS are re-exported from server, and
ToolError/SageProcessError with them, because sagemath_mcp.server is the
documented import surface -- the console script, python -m, the CLI
integration configs and the suite all reach for it. Private helpers were
not re-exported; the handful of tests using them now name their real
home, codegen or tools.core.

Two module-level items had to follow their users rather than stay behind:
_PLOT3D_GRID to plotting, and a logger for core.

Step 5 came along for the ride because it had to: the generated-code lint
resolved src/sagemath_mcp/server.py by path. After the move that file
still existed and still parsed, it just held no tools -- so the caret
lint, the save(_buf) lint, the documented-example check and the item-18
guard would all have passed while inspecting nothing. It scans the whole
package now, and the discovery floor caught this exactly as intended
(assert 0 >= 30) rather than going quiet.

Widening the scan surfaced one real gap in it: the caret check excluded
re.compile patterns but not re.sub/match/search, so session.py's journal
sanitiser looked like XOR in generated code.

Verified: the tool inventory snapshot is unchanged (37 tools, 3
templates, same schemas and descriptions), 357 unit and 433 integration
tests pass, both entry points start, and the item-18 guard was re-proved
after retargeting by reintroducing a raw interpolation.
