+++ src/sagemath_mcp/server.py
@@ -3,44 +3,79 @@
 from __future__ import annotations
 
 import argparse
-import asyncio
-import contextlib
 import logging
-import textwrap
-from typing import Annotated
 
-from fastmcp import Context
-from fastmcp.exceptions import ToolError
-from pydantic import Field
+from fastmcp.exceptions import ToolError  # noqa: F401 - re-exported for callers and tests
 
-from . import __version__, monitoring, runtime
+from . import (
+    __version__,
+    runtime,
+    tools,  # noqa: F401 - imported for its registration side effect
+)
 from .app import mcp
-from .codegen import (
-    _NAMED_GRAPH_RE,
-    _check_matrix,
-    _declare_free_symbols,
-    _distribution_mean,
-    _distribution_variance,
-    _encode_literal,
-    _evaluate_structured,
-    _exact_int,
-    _normal_parameters,
-    _sage_prelude,
-    _validated_expression,
-    _validated_identifier,
-)
-from .config import DEFAULT_SETTINGS
-from .models import (
-    DocumentationLink,
-    EvaluateResult,
-    MonitoringSnapshot,
-    ResetResponse,
-    SessionSnapshot,
-)
+from .config import DEFAULT_SETTINGS  # noqa: F401 - part of this module's long-standing surface
 from .session import (
     DEFAULT_SESSION_NAME,
-    SageEvaluationError,
-    SageProcessError,
+    SageProcessError,  # noqa: F401 - re-exported
+)
+
+# The tool functions are re-exported here because `sagemath_mcp.server` is the
+# documented import surface: the console script, `python -m sagemath_mcp.server`,
+# the CLI integration configs and the test suite all reach for it. Importing the
+# tools package is what registers them; these names keep the old spelling working.
+from .tools.algebra import (  # noqa: F401
+    boolean_algebra_operation,
+    matrix_multiply,
+    matrix_operation,
+    polynomial_ring_operation,
+    solve_equation,
+)
+from .tools.calculus import (  # noqa: F401
+    differentiate_expression,
+    integrate_expression,
+    limit_expression,
+    series_expansion,
+    solve_ode,
+    symbolic_sum,
+    vector_calculus_operation,
+)
+from .tools.core import (  # noqa: F401
+    calculate_expression,
+    evaluate_sage,
+    evaluate_sage_streaming,
+    expand_expression,
+    factor_expression,
+    find_root,
+    simplify_expression,
+)
+from .tools.discrete import (  # noqa: F401
+    coding_theory_operation,
+    combinatorics_operation,
+    elliptic_curve_operation,
+    graph_operation,
+    group_operation,
+    number_theory_operation,
+)
+from .tools.plotting import (  # noqa: F401
+    geometry_operation,
+    plot3d_expression,
+    plot_expression,
+    plot_multi_expression,
+)
+from .tools.session import (  # noqa: F401
+    cancel_sage_session,
+    documentation_resource,
+    interrupt_sage_session,
+    list_sage_sessions,
+    monitoring_resource,
+    reset_sage_session,
+    session_resource,
+    start_sage_session,
+    stop_sage_session,
+)
+from .tools.stats import (  # noqa: F401
+    distribution_operation,
+    statistics_summary,
 )
 
 LOGGER = logging.getLogger(__name__)
@@ -49,21 +84,6 @@ _SESSION_ARG_DESC = (
     "Named workspace to use. Workspaces have independent variables; "
     f"omit for '{DEFAULT_SESSION_NAME}'."
 )
-
-DOC_LINKS: list[DocumentationLink] = [
-    DocumentationLink(
-        title="SageMath Reference Manual",
-        url="https://doc.sagemath.org/html/en/reference",
-        slug="reference",
-        description="Comprehensive API and functionality reference for SageMath.",
-    ),
-    DocumentationLink(
-        title="Sage Tutorial",
-        url="https://doc.sagemath.org/html/en/tutorial",
-        slug="tutorial",
-        description="Gentle introduction to SageMath syntax and workflows.",
-    ),
-]
 
 
 # ---------------------------------------------------------------------------
@@ -85,1768 +105,6 @@ async def health_check(request: object) -> object:
     )
 
 
-# NOTE ON THIS DESCRIPTION. It used to say "use this for anything not covered by
-# the specialized helpers" and then demonstrate fourteen domains that ARE covered
-# by them, including codes.HammingCode(...).minimum_distance(). One line of
-# instruction against fourteen lines of demonstration: measured against the CLI
-# suite, Codex chose evaluate_sage for every single question. The examples below
-# are deliberately restricted to work no dedicated tool performs, so that what
-# this tool shows and what it says agree.
-@mcp.tool(description="""\
-Run arbitrary SageMath code in a persistent session; variables persist across calls.
-
-LAST RESORT. A dedicated tool exists for most tasks and should be preferred: it \
-validates arguments and returns a typed result instead of a repr string. Reach for \
-one of these first:
-
-- calculus: differentiate_expression, integrate_expression, limit_expression, \
-series_expansion, symbolic_sum, solve_ode
-- algebra: solve_equation, simplify_expression, expand_expression, \
-factor_expression, find_root
-- linear algebra: matrix_operation (determinant, inverse, eigenvalues, rank, rref, \
-transpose), matrix_multiply
-- discrete: number_theory_operation, combinatorics_operation (binomial, partitions, \
-catalan, fibonacci, bell), graph_operation, group_operation
-- specialised: elliptic_curve_operation, coding_theory_operation, \
-polynomial_ring_operation, boolean_algebra_operation, geometry_operation, \
-vector_calculus_operation
-- data: statistics_summary, distribution_operation
-- plots: plot_expression, plot3d_expression, plot_multi_expression
-
-Use evaluate_sage only for what those do not cover, for example:
-
-Transforms: laplace(sin(t), t, s); inverse_laplace(1/(s^2+1), s, t)
-Modular arithmetic: Mod(17, 5); power_mod(3, 100, 97)
-Recurrences: desolve_rsolve(f(n+2)-f(n+1)-f(n), f, [0, 1])
-Continued fractions: continued_fraction(pi, nterms=10)
-Number fields: K.<a> = NumberField(x^3 - 2); K.class_number()
-Multi-step work that builds on values defined earlier in the same session.
-""")
-async def evaluate_sage(
-    code: Annotated[str, Field(description="SageMath code to execute")],
-    want_latex: Annotated[
-        bool, Field(description="Return LaTeX representation when possible")
-    ] = False,
-    capture_stdout: Annotated[
-        bool, Field(description="Capture stdout emitted by Sage code")
-    ] = True,
-    timeout_seconds: Annotated[
-        float | None,
-        Field(
-            description="Override the evaluation timeout in seconds",
-            alias="timeout",
-            validation_alias="timeout",
-            serialization_alias="timeout",
-            gt=0.0,
-            default=None,
-        ),
-    ] = None,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> EvaluateResult:
-    """Run SageMath code, preserving state within the caller's MCP session."""
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    # Compute the key once and reuse it. Cancelling used to pass ctx.session_id,
-    # which restarts the DEFAULT workspace: cancelling work in 'curves' destroyed
-    # unrelated default state while the curves worker kept running.
-    session_key = runtime.SESSION_MANAGER.key_for(ctx.session_id, session)
-    sage_session = await runtime.SESSION_MANAGER.get(session_key)
-    progress_task: asyncio.Task[None] | None = None
-    if ctx is not None:
-        await ctx.info("Starting SageMath evaluation")
-        progress_task = asyncio.create_task(_progress_heartbeat(ctx))
-    try:
-        worker_result = await sage_session.evaluate(
-            code,
-            want_latex=want_latex,
-            capture_stdout=capture_stdout,
-            timeout_seconds=timeout_seconds,
-        )
-    except asyncio.CancelledError:
-        monitoring.record_failure("cancelled", is_security=False, details="evaluation cancelled")
-        await runtime.SESSION_MANAGER.cancel(session_key)
-        if ctx is not None:
-            await ctx.warning(f"Sage evaluation cancelled; session '{session}' restarted")
-        raise
-    except SageEvaluationError as exc:
-        monitoring.record_failure(
-            exc.error_type or str(exc),
-            is_security=exc.error_type == "SecurityViolation",
-            details=exc.traceback or exc.stdout,
-        )
-        if ctx is not None:
-            if exc.error_type == "SecurityViolation":
-                await ctx.error(f"Sage security policy violation: {exc}")
-            else:
-                await ctx.error(f"SageMath error: {exc}")
-        raise ToolError(exc.args[0]) from exc
-    except SageProcessError as exc:
-        cause = getattr(exc, "__cause__", None)
-        details = repr(cause) if cause is not None else None
-        monitoring.record_failure(
-            str(exc) or exc.__class__.__name__,
-            is_security=False,
-            details=details,
-        )
-        if ctx is not None:
-            await ctx.error("SageMath process became unavailable; restarting may help")
-        raise ToolError(str(exc)) from exc
-    finally:
-        if progress_task:
-            progress_task.cancel()
-            with contextlib.suppress(asyncio.CancelledError):
-                await progress_task
-        if ctx is not None:
-            await ctx.report_progress(1.0, 1.0, "Sage evaluation complete")
-    monitoring.record_success(worker_result.elapsed_ms)
-    return EvaluateResult(
-        result_type=worker_result.result_type,
-        result=worker_result.result,
-        latex=worker_result.latex,
-        stdout=_truncate_stdout(worker_result.stdout),
-        elapsed_ms=worker_result.elapsed_ms,
-    )
-
-
-@mcp.tool(description="Reset the SageMath session state for the current MCP session")
-async def reset_sage_session(
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> ResetResponse:
-    """Reset the Sage session associated with the current MCP session."""
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required to reset state")
-    await runtime.SESSION_MANAGER.reset(runtime.SESSION_MANAGER.key_for(ctx.session_id, session))
-    await ctx.info(f"Sage session '{session}' reset")
-    return ResetResponse()
-
-
-@mcp.tool(
-    description="Interrupt a running Sage computation while keeping variables defined so far"
-)
-async def interrupt_sage_session(
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> ResetResponse:
-    """Signal the worker to abandon its computation without losing state.
-
-    Prefer this over cancel_sage_session: cancelling restarts the worker and
-    discards every variable, which is the worse outcome when the state was
-    expensive to build.
-    """
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required to interrupt work")
-    key = runtime.SESSION_MANAGER.key_for(ctx.session_id, session)
-    interrupted = await runtime.SESSION_MANAGER.interrupt(key)
-    if not interrupted:
-        # No worker to signal: either nothing has run yet in this workspace, or
-        # it has already exited. Not an error, but say which.
-        await ctx.info(f"No running Sage worker for session '{session}'")
-        return ResetResponse(message=f"No running computation in session '{session}'")
-    await ctx.warning(f"Interrupted session '{session}'; state preserved")
-    return ResetResponse(message=f"Interrupted session '{session}'; state preserved")
-
-
-@mcp.tool(description="Cancel any running Sage computation and restart the worker")
-async def cancel_sage_session(
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> ResetResponse:
-    """Cancel in-flight work by restarting the backing Sage worker.
-
-    This discards the namespace. Use interrupt_sage_session to stop a
-    computation while keeping it.
-    """
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required to cancel work")
-    await runtime.SESSION_MANAGER.cancel(runtime.SESSION_MANAGER.key_for(ctx.session_id, session))
-    await ctx.warning(f"Sage session '{session}' cancelled and restarted")
-    return ResetResponse(message="Session cancelled and restarted")
-
-
-@mcp.tool(description="Start a named Sage workspace with its own independent variables")
-async def start_sage_session(
-    name: Annotated[str, Field(description="Workspace name, e.g. 'curves' or 'scratch'")],
-    ctx: Context | None = None,
-) -> ResetResponse:
-    """Create a named workspace so one client can hold several at once.
-
-    Workspaces are independent: a variable defined in one is invisible to the
-    others. Calling this for an existing name is harmless.
-    """
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required to start a session")
-    if not name.strip():
-        raise ToolError("Session name must not be empty")
-    await runtime.resolve_session(ctx.session_id, name)
-    await ctx.info(f"Started Sage session '{name}'")
-    return ResetResponse(message=f"Session '{name}' ready")
-
-
-@mcp.tool(description="List the named Sage workspaces belonging to this client")
-async def list_sage_sessions(ctx: Context | None = None) -> dict:
-    """Report every workspace for this client, with liveness and statement counts."""
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required to list sessions")
-    sessions = await runtime.SESSION_MANAGER.list_for_scope(ctx.session_id)
-    return {"sessions": sessions, "count": len(sessions)}
-
-
-@mcp.tool(description="Stop a named Sage workspace and release its worker")
-async def stop_sage_session(
-    name: Annotated[str, Field(description="Workspace name to stop")],
-    ctx: Context | None = None,
-) -> ResetResponse:
-    """Terminate one workspace. Other workspaces are unaffected."""
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required to stop a session")
-    stopped = await runtime.SESSION_MANAGER.stop(ctx.session_id, name)
-    if not stopped:
-        raise ToolError(f"No Sage session named '{name}' for this client")
-    await ctx.info(f"Stopped Sage session '{name}'")
-    return ResetResponse(message=f"Session '{name}' stopped")
-
-
-@mcp.resource("resource://sagemath/session/{scope}")
-async def session_resource(scope: str, ctx: Context | None = None) -> str:
-    """Expose a resource describing active Sage sessions for observability."""
-    import json as _json
-
-    del ctx  # resource does not require request context
-    data = runtime.SESSION_MANAGER.snapshot()
-    if scope != "all":
-        data = [entry for entry in data if entry["session_id"] == scope]
-    snapshots = [
-        SessionSnapshot(
-            session_id=entry["session_id"],
-            live=bool(entry["live"]),
-            started_at=float(entry["started_at"]),
-            last_used_at=float(entry["last_used_at"]),
-            idle_seconds=float(entry["idle_seconds"]),
-        )
-        for entry in data
-    ]
-    return _json.dumps([s.model_dump() for s in snapshots])
-
-
-@mcp.resource("resource://sagemath/monitoring/{scope}")
-async def monitoring_resource(scope: str, ctx: Context | None = None) -> str:
-    """Expose aggregated metrics for observability."""
-    del ctx
-    if scope not in {"metrics", "all"}:
-        return "[]"
-    snapshot = monitoring.snapshot()
-    return MonitoringSnapshot(**snapshot).model_dump_json()
-
-
-async def _progress_heartbeat(ctx: Context, interval: float = 1.5) -> None:
-    """Emit periodic progress updates so clients can show activity."""
-    elapsed = 0.0
-    try:
-        while True:
-            await asyncio.sleep(interval)
-            elapsed += interval
-            await ctx.report_progress(elapsed, None, f"Sage running for {elapsed:.1f}s")
-    except asyncio.CancelledError:  # pragma: no cover - background task shutdown
-        return
-
-
-def _truncate_stdout(stdout: str) -> str:
-    """Clamp stdout to the configured limit while signalling truncation."""
-    settings = runtime.get_session_manager().settings
-    limit = getattr(settings, "max_stdout_chars", DEFAULT_SETTINGS.max_stdout_chars)
-    if not isinstance(limit, int):  # defensive: shared settings may be class-level descriptors
-        limit = DEFAULT_SETTINGS.max_stdout_chars
-    if len(stdout) <= limit:
-        return stdout
-    clipped = stdout[:limit]
-    LOGGER.warning(
-        "Truncated Sage stdout to %s characters (requested %s)", limit, len(stdout)
-    )
-    return clipped + "\n… [output truncated]"
-
-
-@mcp.resource("resource://sagemath/docs/{scope}")
-async def documentation_resource(scope: str, ctx: Context | None = None) -> list[DocumentationLink]:
-    del ctx
-    if scope == "all":
-        return DOC_LINKS
-    return [link for link in DOC_LINKS if link.slug == scope]
-
-
-# Samples per axis for the 3D surface. 48x48 keeps the rendered surface smooth
-# while staying well inside the evaluation timeout.
-_PLOT3D_GRID = 48
-
-
-@mcp.tool(description="Evaluate a SageMath expression and return numeric/string forms")
-async def calculate_expression(
-    expression: Annotated[str, Field(description="SageMath expression to evaluate")],
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = (
-        _sage_prelude()
-        + textwrap.dedent(
-            f"""
-        _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-        if hasattr(_expr, 'n'):
-            try:
-                _numeric = float(_expr.n())
-            except (TypeError, ValueError):
-                _numeric = None
-        else:
-            try:
-                _numeric = float(_expr)
-            except (TypeError, ValueError):
-                _numeric = None
-        _payload = {{'string': str(_expr)}}
-        if _numeric is not None:
-            _payload['numeric'] = _numeric
-        _payload
-        """
-        )
-    )
-    payload = await _evaluate_structured(session, code)
-    if not isinstance(payload, dict):
-        return {"string": str(payload)}
-    return payload
-
-
-@mcp.tool(description="Solve an equation or system of equations")
-async def solve_equation(
-    equation: Annotated[
-        str | list[str],
-        Field(description="Equation string (e.g., 'x^2 - 1 = 0') or list of equations for systems"),
-    ],
-    variable: Annotated[
-        str | list[str],
-        Field(description="Variable or list of variables to solve for", default="x"),
-    ] = "x",
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    equations = [equation] if isinstance(equation, str) else equation
-    variables = [variable] if isinstance(variable, str) else variable
-    code = (
-        _sage_prelude(variables)
-        + textwrap.dedent(
-            f"""
-        _vars = [var(v) for v in {_encode_literal(variables)}]
-        _eqs = []
-        for _eq_str in {_encode_literal(equations)}:
-            parts = _eq_str.split('=')
-            if len(parts) == 2:
-                left = sage_eval(parts[0].strip(), locals=_locals)
-                right = sage_eval(parts[1].strip(), locals=_locals)
-                _eqs.append(left == right)
-            else:
-                _eqs.append(sage_eval(_eq_str, locals=_locals))
-        if len(_eqs) == 1 and len(_vars) == 1:
-            _solutions = solve(_eqs[0], _vars[0])
-        else:
-            _solutions = solve(_eqs, _vars)
-        [str(sol) for sol in _solutions]
-        """
-        )
-    )
-    solutions = await _evaluate_structured(session, code)
-    return {"solutions": solutions}
-
-
-@mcp.tool(description="Differentiate an expression with respect to a variable")
-async def differentiate_expression(
-    expression: Annotated[str, Field(description="Expression to differentiate")],
-    variable: Annotated[str, Field(description="Variable for differentiation", default="x")] = "x",
-    order: Annotated[
-        int,
-        Field(description="Order of differentiation (1 = first, 2 = second, etc.)", ge=1),
-    ] = 1,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = (
-        _sage_prelude([variable])
-        + textwrap.dedent(
-            f"""
-        _var = var({_encode_literal(variable)})
-        _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-        str(diff(_expr, _var, {order}))
-        """
-        )
-    )
-    result = await _evaluate_structured(session, code)
-    return {"derivative": result, "order": order}
-
-
-@mcp.tool(description="Integrate an expression (indefinite or definite with bounds)")
-async def integrate_expression(
-    expression: Annotated[str, Field(description="Expression to integrate")],
-    variable: Annotated[str, Field(description="Integration variable", default="x")] = "x",
-    lower_bound: Annotated[
-        str | None,
-        Field(description="Lower bound for definite integral (e.g., '0', '-oo')"),
-    ] = None,
-    upper_bound: Annotated[
-        str | None,
-        Field(description="Upper bound for definite integral (e.g., '1', 'oo')"),
-    ] = None,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    if (lower_bound is None) != (upper_bound is None):
-        raise ToolError("Both lower_bound and upper_bound must be provided for a definite integral")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    definite = lower_bound is not None
-    if definite:
-        code = (
-            _sage_prelude([variable])
-            + textwrap.dedent(
-                f"""
-            _var = var({_encode_literal(variable)})
-            _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-            {_declare_free_symbols(lower_bound, upper_bound)}
-            _lb = sage_eval({_encode_literal(lower_bound)}, locals=_locals)
-            _ub = sage_eval({_encode_literal(upper_bound)}, locals=_locals)
-            str(integrate(_expr, _var, _lb, _ub))
-            """
-            )
-        )
-    else:
-        code = (
-            _sage_prelude([variable])
-            + textwrap.dedent(
-                f"""
-            _var = var({_encode_literal(variable)})
-            _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-            str(integrate(_expr, _var))
-            """
-            )
-        )
-    result = await _evaluate_structured(session, code)
-    return {"integral": result, "definite": definite}
-
-
-@mcp.tool(description=(
-        "Descriptive statistics for a list of numbers: mean, median, population "
-        "and sample variance and standard deviation, min and max. Prefer this "
-        "over evaluate_sage for summary statistics."
-    ))
-async def statistics_summary(
-    data: Annotated[list[float], Field(description="List of numeric values")],
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    # Without this the generated code raised a bare "list index out of range"
-    # from the median calculation, which says nothing about what to send instead.
-    if not data:
-        raise ToolError("statistics_summary requires at least one value in 'data'")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = (
-        _sage_prelude()
-        + textwrap.dedent(
-            f"""
-        _data = {_encode_literal(data)}
-        _n = len(_data)
-        _mean = float(mean(_data))
-        _sorted = sorted(_data)
-        _mid = _n // 2
-        _median = float((_sorted[_mid] + _sorted[~_mid]) / 2)
-        _pvar = float(sum((x - _mean)**2 for x in _data) / _n)
-        _svar = float(sum((x - _mean)**2 for x in _data) / (_n - 1)) if _n > 1 else 0.0
-        {{
-            'mean': _mean,
-            'median': _median,
-            'population_variance': _pvar,
-            'sample_variance': _svar,
-            'population_std_dev': float(sqrt(_pvar)),
-            'sample_std_dev': float(sqrt(_svar)),
-            'min': float(min(_data)),
-            'max': float(max(_data)),
-        }}
-        """
-        )
-    )
-    return await _evaluate_structured(session, code)
-
-
-@mcp.tool(description="Multiply two matrices and return the result as nested lists")
-async def matrix_multiply(
-    matrix_a: Annotated[list[list[float]], Field(description="Left matrix (rows of numbers)")],
-    matrix_b: Annotated[list[list[float]], Field(description="Right matrix (rows of numbers)")],
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    # Checked here so a shape mismatch reports the shapes. Left to Sage it
-    # surfaced as "unsupported operand parent(s) for *: 'Full MatrixSpace of
-    # ...'", which does not say which dimension is wrong.
-    _check_matrix(matrix_a, "matrix_a")
-    _check_matrix(matrix_b, "matrix_b")
-    if len(matrix_a[0]) != len(matrix_b):
-        raise ToolError(
-            f"Cannot multiply a {len(matrix_a)}x{len(matrix_a[0])} matrix by a "
-            f"{len(matrix_b)}x{len(matrix_b[0])} matrix: the number of columns in "
-            "matrix_a must equal the number of rows in matrix_b"
-        )
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = textwrap.dedent(
-        f"""
-        from sage.all import *
-        A = matrix(SR, {matrix_a})
-        B = matrix(SR, {matrix_b})
-        C = A * B
-        [[float(entry) if entry in RR else str(entry) for entry in row] for row in C.rows()]
-        """
-    )
-    product = await _evaluate_structured(session, code)
-    return {"product": product}
-
-
-@mcp.tool(description="Simplify a mathematical expression")
-async def simplify_expression(
-    expression: Annotated[str, Field(description="Expression to simplify")],
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = (
-        _sage_prelude()
-        + textwrap.dedent(
-            f"""
-        _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-        str(simplify(_expr))
-        """
-        )
-    )
-    result = await _evaluate_structured(session, code)
-    return {"simplified": result}
-
-
-@mcp.tool(description="Expand a mathematical expression")
-async def expand_expression(
-    expression: Annotated[str, Field(description="Expression to expand")],
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = (
-        _sage_prelude()
-        + textwrap.dedent(
-            f"""
-        _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-        str(expand(_expr))
-        """
-        )
-    )
-    result = await _evaluate_structured(session, code)
-    return {"expanded": result}
-
-
-@mcp.tool(description="Factor a mathematical expression or integer")
-async def factor_expression(
-    expression: Annotated[str, Field(description="Expression to factor (e.g., 'x^2 - 1' or '60')")],
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = (
-        _sage_prelude()
-        + textwrap.dedent(
-            f"""
-        _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-        str(factor(_expr))
-        """
-        )
-    )
-    result = await _evaluate_structured(session, code)
-    return {"factored": result}
-
-
-@mcp.tool(description="Compute the limit of an expression")
-async def limit_expression(
-    expression: Annotated[str, Field(description="Expression to take the limit of")],
-    variable: Annotated[str, Field(description="Variable approaching the point")] = "x",
-    point: Annotated[str, Field(description="Point to approach (e.g., '0', 'oo', '-oo')")] = "0",
-    direction: Annotated[
-        str | None,
-        Field(description="Direction: 'plus' (right), 'minus' (left), or omit for both"),
-    ] = None,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    dir_arg = f", dir={_encode_literal(direction)}" if direction else ""
-    code = (
-        _sage_prelude([variable])
-        + textwrap.dedent(
-            f"""
-        _var = var({_encode_literal(variable)})
-        _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-        {_declare_free_symbols(point)}
-        _point = sage_eval({_encode_literal(point)}, locals=_locals)
-        str(limit(_expr, _var, _point{dir_arg}))
-        """
-        )
-    )
-    result = await _evaluate_structured(session, code)
-    return {"limit": result}
-
-
-@mcp.tool(description="Compute a Taylor/Laurent series expansion")
-async def series_expansion(
-    expression: Annotated[str, Field(description="Expression to expand in series")],
-    variable: Annotated[str, Field(description="Variable for expansion")] = "x",
-    point: Annotated[str, Field(description="Point around which to expand")] = "0",
-    order: Annotated[int, Field(description="Number of terms in the expansion", ge=1)] = 6,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = (
-        _sage_prelude([variable])
-        + textwrap.dedent(
-            f"""
-        _var = var({_encode_literal(variable)})
-        _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-        {_declare_free_symbols(point)}
-        _point = sage_eval({_encode_literal(point)}, locals=_locals)
-        str(_expr.series(_var == _point, {order}))
-        """
-        )
-    )
-    result = await _evaluate_structured(session, code)
-    return {"series": result, "point": point, "order": order}
-
-
-@mcp.tool(description=(
-        "Linear algebra on one matrix: determinant, inverse, eigenvalues, rank, "
-        "reduced row echelon form, transpose. Prefer this over evaluate_sage."
-    ))
-async def matrix_operation(
-    matrix: Annotated[
-        list[list[float]], Field(description="Matrix as nested list of numbers")
-    ],
-    operation: Annotated[
-        str,
-        Field(description="One of: determinant, inverse, eigenvalues, rank, rref, transpose"),
-    ],
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    operation = operation.strip()
-    _check_matrix(matrix, "matrix")
-    allowed_ops = {"determinant", "inverse", "eigenvalues", "rank", "rref", "transpose"}
-    if operation not in allowed_ops:
-        raise ToolError(
-            f"Unknown operation '{operation}'. "
-            f"Must be one of: {', '.join(sorted(allowed_ops))}"
-        )
-    session = await runtime.resolve_session(ctx.session_id, session)
-    _row_repr = (
-        "[[float(e) if e in RR else str(e) for e in row] for row in {obj}.rows()]"
-    )
-    op_code = {
-        "determinant": (
-            "float(M.determinant()) if M.determinant() in RR"
-            " else str(M.determinant())"
-        ),
-        "inverse": _row_repr.format(obj="M.inverse()"),
-        "eigenvalues": (
-            "[float(ev) if ev in RR else str(ev) for ev in M.eigenvalues()]"
-        ),
-        "rank": "int(M.rank())",
-        "rref": _row_repr.format(obj="M.rref()"),
-        "transpose": _row_repr.format(obj="M.transpose()"),
-    }
-    code = textwrap.dedent(
-        f"""
-        from sage.all import *
-        M = matrix(SR, {matrix})
-        {op_code[operation]}
-        """
-    )
-    result = await _evaluate_structured(session, code)
-    return {"operation": operation, "result": result}
-
-
-@mcp.tool(description=(
-        "Solve an ordinary differential equation of any order, returning the "
-        "general solution with arbitrary constants. Prefer this over evaluate_sage."
-    ))
-async def solve_ode(
-    equation: Annotated[
-        str,
-        Field(description="ODE string, e.g., \"diff(y(x),x) + y(x) = 0\""),
-    ],
-    function: Annotated[str, Field(description="Dependent function name (e.g., 'y')")] = "y",
-    variable: Annotated[str, Field(description="Independent variable (e.g., 'x')")] = "x",
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = (
-        _sage_prelude([variable])
-        + textwrap.dedent(
-            f"""
-        _x = var({_encode_literal(variable)})
-        _ode_function = function({_encode_literal(function)})
-        _y = _ode_function(_x)
-        _ode_text = {_encode_literal(equation)}
-
-        def _build_ode(_binding):
-            _ode_locals = dict(_locals)
-            _ode_locals[{_encode_literal(function)}] = _binding
-            _ode_locals['diff'] = diff
-            parts = _ode_text.split('=')
-            if len(parts) == 2:
-                left = sage_eval(parts[0].strip(), locals=_ode_locals)
-                right = sage_eval(parts[1].strip(), locals=_ode_locals)
-                return left == right
-            return sage_eval(_ode_text, locals=_ode_locals)
-
-        # Bind the bare name to the undefined function so the documented
-        # "diff(y(x), x)" form parses. Binding the applied expression instead
-        # turns "y(x)" into "(y(x))(x)", which Sage rejects with "Substitution
-        # using function-call syntax and unnamed arguments has been removed".
-        # Fall back to the applied expression so a bare "diff(y, x)" still
-        # works, since that form cannot be parsed against the function itself.
-        try:
-            _ode = _build_ode(_ode_function)
-        except Exception:
-            _ode = _build_ode(_y)
-        str(desolve(_ode, _y, ivar=_x))
-        """
-        )
-    )
-    result = await _evaluate_structured(session, code)
-    return {"solution": result}
-
-
-@mcp.tool(description=(
-        "Number theory: primality testing, integer factorisation, the next "
-        "prime above n, gcd and lcm. Prefer this over evaluate_sage for any of these."
-    ))
-async def number_theory_operation(
-    operation: Annotated[
-        str,
-        Field(description="Operation: 'is_prime', 'factor_integer', 'next_prime', 'gcd', 'lcm'"),
-    ],
-    a: Annotated[
-        int | str,
-        Field(
-            description=(
-                "Primary integer. Pass values above 2^53 as a decimal STRING: "
-                "JSON numbers are IEEE doubles in JavaScript-based clients, so "
-                "10^30 arrives as 1000000000000000019884624838656 and the answer "
-                "is silently wrong."
-            )
-        ),
-    ],
-    b: Annotated[
-        int | str | None,
-        Field(description="Second integer, required for gcd and lcm. Same string rule."),
-    ] = None,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    operation = operation.strip()
-    a = _exact_int(a, "a")
-    b = _exact_int(b, "b") if b is not None else None
-    allowed_ops = {"is_prime", "factor_integer", "next_prime", "gcd", "lcm"}
-    if operation not in allowed_ops:
-        raise ToolError(
-            f"Unknown operation '{operation}'. "
-            f"Must be one of: {', '.join(sorted(allowed_ops))}"
-        )
-    if operation in {"gcd", "lcm"} and b is None:
-        raise ToolError(f"Operation '{operation}' requires both 'a' and 'b' arguments")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    op_code = {
-        "is_prime": f"bool(is_prime({a}))",
-        "factor_integer": f"str(factor({a}))",
-        "next_prime": f"int(next_prime({a}))",
-        "gcd": f"int(gcd({a}, {b}))",
-        "lcm": f"int(lcm({a}, {b}))",
-    }
-    code = _sage_prelude() + op_code[operation] + "\n"
-    result = await _evaluate_structured(session, code)
-    return {"operation": operation, "result": result}
-
-
-@mcp.tool(description=(
-        "Closed form of a symbolic sum or product over an index variable, "
-        "including infinite series. Prefer this over evaluate_sage for summations."
-    ))
-async def symbolic_sum(
-    expression: Annotated[str, Field(description="Expression to sum (e.g. '1/n^2')")],
-    variable: Annotated[str, Field(description="Index variable (e.g. 'n')")] = "n",
-    lower: Annotated[str, Field(description="Lower bound (e.g. '1')")] = "1",
-    upper: Annotated[str, Field(description="Upper bound (e.g. 'oo' for infinity)")] = "oo",
-    product: Annotated[
-        bool, Field(description="If true, compute a product instead of a sum")
-    ] = False,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    op = "product" if product else "sum"
-    code = (
-        _sage_prelude([variable])
-        + textwrap.dedent(
-            f"""
-        _var = var({_encode_literal(variable)})
-        _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-        {_declare_free_symbols(lower, upper)}
-        _lo = sage_eval({_encode_literal(lower)}, locals=_locals)
-        _hi = sage_eval({_encode_literal(upper)}, locals=_locals)
-        str({op}(_expr, _var, _lo, _hi))
-        """
-        )
-    )
-    result = await _evaluate_structured(session, code)
-    return {"result": result, "operation": op}
-
-
-@mcp.tool(description=(
-        "Combinatorics: binomial coefficients, permutations, combinations, "
-        "integer partitions, factorial, Catalan, Fibonacci and Bell numbers. "
-        "Prefer this over evaluate_sage for any of these."
-    ))
-async def combinatorics_operation(
-    operation: Annotated[
-        str,
-        Field(
-            # Every entry says what it returns. "partitions" alone left it
-            # ambiguous whether the result was a count or a list of partitions,
-            # and a client asking "how many partitions does 120 have" reached
-            # for evaluate_sage rather than risk the wrong shape.
-            description="One of: binomial (n choose k), permutations (n!), "
-            "combinations (n choose k), partitions (COUNT of integer partitions "
-            "of n), factorial (n!), catalan (nth Catalan number), fibonacci "
-            "(nth Fibonacci number), bell (nth Bell number). All return a single "
-            "integer."
-        ),
-    ],
-    n: Annotated[int, Field(description="Primary integer argument")],
-    k: Annotated[
-        int | None, Field(description="Secondary argument (for binomial, combinations)")
-    ] = None,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    operation = operation.strip()
-    session = await runtime.resolve_session(ctx.session_id, session)
-    op_code = {
-        "binomial": f"int(binomial({n}, {k or 0}))",
-        "permutations": f"int(Permutations({n}).cardinality())"
-        if k is None
-        else f"int(factorial({n}) // factorial({n} - {k}))",
-        "combinations": f"int(binomial({n}, {k or 0}))",
-        "partitions": f"int(Partitions({n}).cardinality())",
-        "factorial": f"int(factorial({n}))",
-        "catalan": f"int(catalan_number({n}))",
-        "fibonacci": f"int(fibonacci({n}))",
-        "bell": f"int(bell_number({n}))",
-    }
-    if operation not in op_code:
-        raise ToolError(f"Unknown operation '{operation}'. Use: {', '.join(op_code)}")
-    code = _sage_prelude() + op_code[operation] + "\n"
-    result = await _evaluate_structured(session, code)
-    return {"operation": operation, "result": result}
-
-
-@mcp.tool(description="Plot a 3D surface of a two-variable expression as base64 PNG")
-async def plot3d_expression(
-    expression: Annotated[
-        str, Field(description="Expression of two variables (e.g. 'sin(x)*cos(y)')")
-    ],
-    x_variable: Annotated[str, Field(description="First variable")] = "x",
-    y_variable: Annotated[str, Field(description="Second variable")] = "y",
-    x_range_min: Annotated[float, Field(description="X lower bound")] = -5.0,
-    x_range_max: Annotated[float, Field(description="X upper bound")] = 5.0,
-    y_range_min: Annotated[float, Field(description="Y lower bound")] = -5.0,
-    y_range_max: Annotated[float, Field(description="Y upper bound")] = 5.0,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = (
-        _sage_prelude([x_variable, y_variable])
-        + textwrap.dedent(
-            f"""
-        import base64
-        import io as _io
-        from sage.plot.graphics import Graphics as _Graphics
-        _xv = var({_encode_literal(x_variable)})
-        _yv = var({_encode_literal(y_variable)})
-        _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-        # Sage's plot3d returns a Graphics3d, whose save()/save_image() require
-        # a filesystem path and reject a BytesIO. There is no .matplotlib()
-        # figure on it either, and a temp file is unreachable from the sandbox
-        # (`open` is forbidden, tempfile/os are not importable). So sample the
-        # surface and render it through matplotlib's 3D axes, which writes to
-        # memory. A 2D Graphics is only used to obtain a Figure without
-        # importing matplotlib directly.
-        try:
-            _f = fast_callable(_expr, vars=(_xv, _yv), domain=float)
-        except Exception:
-            _f = None
-
-        def _z_at(_a, _b):
-            # Singular or complex-valued points become NaN, which matplotlib
-            # renders as a gap rather than failing the whole plot.
-            try:
-                if _f is not None:
-                    return float(_f(_a, _b))
-                return float(_expr.subs({{_xv: _a, _yv: _b}}))
-            except Exception:
-                return float('nan')
-
-        _n = {_PLOT3D_GRID}
-        _xlo, _xhi = float({x_range_min}), float({x_range_max})
-        _ylo, _yhi = float({y_range_min}), float({y_range_max})
-        _gx, _gy, _gz = [], [], []
-        for _i in range(_n):
-            _a = _xlo + (_xhi - _xlo) * _i / (_n - 1)
-            for _j in range(_n):
-                _b = _ylo + (_yhi - _ylo) * _j / (_n - 1)
-                _gx.append(_a)
-                _gy.append(_b)
-                _gz.append(_z_at(_a, _b))
-        _fig = _Graphics().matplotlib()
-        _fig.clf()
-        _ax = _fig.add_subplot(111, projection='3d')
-        # plot_trisurf accepts flat sequences, so no numpy import is needed.
-        _ax.plot_trisurf(_gx, _gy, _gz, cmap='viridis')
-        _ax.set_xlabel({_encode_literal(x_variable)})
-        _ax.set_ylabel({_encode_literal(y_variable)})
-        _buf = _io.BytesIO()
-        _fig.savefig(_buf, format='png')
-        _buf.seek(0)
-        base64.b64encode(_buf.read()).decode('ascii')
-        """
-        )
-    )
-    result = await _evaluate_structured(session, code)
-    return {"image_base64": result, "format": "png"}
-
-
-@mcp.tool(
-    description="Probability distribution operations: PDF, CDF, quantile, mean, variance, sampling"
-)
-async def distribution_operation(
-    distribution: Annotated[
-        str,
-        Field(
-            description="Distribution name: normal, exponential, poisson, "
-            "chi_squared, student_t, uniform, beta, gamma"
-        ),
-    ],
-    parameters: Annotated[
-        list[float], Field(description="Distribution parameters (e.g. [0, 1] for standard normal)")
-    ],
-    operation: Annotated[
-        str, Field(description="One of: pdf, cdf, quantile, mean, variance, sample")
-    ],
-    x: Annotated[float | None, Field(description="Point for pdf/cdf/quantile evaluation")] = None,
-    n: Annotated[int | None, Field(description="Number of samples (for sample operation)")] = None,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    operation = operation.strip()
-    session = await runtime.resolve_session(ctx.session_id, session)
-    params_str = ", ".join(str(p) for p in parameters)
-    # "normal" takes [mu, sigma]. The previous mapping passed parameters[0] as
-    # sigma only when exactly one parameter was given and otherwise hardcoded
-    # 1, so [0, 3] silently computed with sigma=1, and mu was never applied at
-    # all. Sage's gaussian is always centred on 0, so mu is applied by shifting
-    # the evaluation point.
-    normal_mu, normal_sigma = _normal_parameters(parameters)
-    dist_map = {
-        "normal": f"RealDistribution('gaussian', {normal_sigma})",
-        "exponential": f"RealDistribution('exponential', {parameters[0] if parameters else 1})",
-        "uniform": f"RealDistribution('uniform', [{params_str}])",
-        "chi_squared": f"RealDistribution('chisquared', {parameters[0] if parameters else 1})",
-        "student_t": f"RealDistribution('t', {parameters[0] if parameters else 1})",
-        "beta": f"RealDistribution('beta', [{params_str}])",
-        "gamma": f"RealDistribution('gamma', [{params_str}])",
-    }
-    # For distributions not directly in RealDistribution, use scipy-like Sage constructs
-    if distribution == "poisson":
-        # Poisson is discrete; handle separately
-        lam = parameters[0] if parameters else 1
-        op_code = {
-            "pdf": (
-                f"float(exp(-{lam}) * {lam}**{x} / factorial(int({x})))"
-                if x is not None else "0"
-            ),
-            "cdf": (
-                f"float(sum(exp(-{lam}) * {lam}**k / factorial(k)"
-                f" for k in range(int({x}) + 1)))"
-                if x is not None else "0"
-            ),
-            "mean": f"float({lam})",
-            "variance": f"float({lam})",
-            "sample": f"[int(numpy_rng.poisson({lam})) for _ in range({n or 1})]",
-        }
-        if operation not in op_code:
-            raise ToolError(f"Unknown operation '{operation}' for Poisson distribution")
-        code = _sage_prelude() + op_code.get(operation, "None") + "\n"
-    elif distribution in dist_map:
-        dist_expr = dist_map[distribution]
-        # Only the normal distribution carries a location parameter here; for
-        # every other distribution the shift is 0 and these read unchanged.
-        shift = normal_mu if distribution == "normal" else 0.0
-        shifted = None if x is None else f"({x}) - ({shift})"
-        unshift = f"({shift}) + " if distribution == "normal" else ""
-        op_code = {
-            "pdf": f"float(_d.distribution_function({shifted}))" if x is not None else "None",
-            "cdf": (
-                f"float(_d.cum_distribution_function({shifted}))"
-                if x is not None else "None"
-            ),
-            "quantile": (
-                f"float({unshift}_d.cum_distribution_function_inv({x}))"
-                if x is not None else "None"
-            ),
-            # mean/variance are computed analytically. They previously
-            # returned float(_d.get_random_element()) and None respectively,
-            # so "mean" reported a random draw from the distribution -- a
-            # different wrong answer on every call -- and "variance" was
-            # always null.
-            "mean": f"float({_distribution_mean(distribution, parameters)})",
-            "variance": f"float({_distribution_variance(distribution, parameters)})",
-            "sample": f"[float(_d.get_random_element()) for _ in range({n or 1})]",
-        }
-        if operation not in op_code:
-            raise ToolError(
-                f"Unknown operation '{operation}'. "
-                "Use: pdf, cdf, quantile, mean, variance, sample"
-            )
-        code = _sage_prelude() + f"_d = {dist_expr}\n" + op_code[operation] + "\n"
-    else:
-        raise ToolError(
-            f"Unknown distribution '{distribution}'. "
-            "Use: normal, exponential, poisson, chi_squared, student_t, uniform, beta, gamma"
-        )
-    result = await _evaluate_structured(session, code)
-    return {"distribution": distribution, "operation": operation, "result": result}
-
-
-@mcp.tool(description="Find a numeric root of an expression in a given interval")
-async def find_root(
-    expression: Annotated[str, Field(description="Expression to find root of (e.g. 'x - cos(x)')")],
-    variable: Annotated[str, Field(description="Variable")] = "x",
-    lower_bound: Annotated[float, Field(description="Left bound of search interval")] = -10.0,
-    upper_bound: Annotated[float, Field(description="Right bound of search interval")] = 10.0,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = (
-        _sage_prelude([variable])
-        + textwrap.dedent(
-            f"""
-        _var = var({_encode_literal(variable)})
-        _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-        float(find_root(_expr, {lower_bound}, {upper_bound}))
-        """
-        )
-    )
-    result = await _evaluate_structured(session, code)
-    return {"root": result}
-
-
-@mcp.tool(description="Plot multiple expressions overlaid on a single 2D graph")
-async def plot_multi_expression(
-    expressions: Annotated[
-        list[str], Field(description="List of expressions to plot (e.g. ['sin(x)', 'cos(x)'])")
-    ],
-    variable: Annotated[str, Field(description="Plot variable")] = "x",
-    range_min: Annotated[float, Field(description="Lower bound of plot range")] = -10.0,
-    range_max: Annotated[float, Field(description="Upper bound of plot range")] = 10.0,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = (
-        _sage_prelude([variable])
-        + textwrap.dedent(
-            f"""
-        import base64
-        import io as _io
-        _var = var({_encode_literal(variable)})
-        _exprs = [sage_eval(e, locals=_locals) for e in {_encode_literal(expressions)}]
-        _plt = sum(plot(e, (_var, {range_min}, {range_max})) for e in _exprs)
-        _buf = _io.BytesIO()
-        # Graphics.save() needs a filesystem path and rejects a BytesIO with
-        # "expected str, bytes or os.PathLike object". Going through the
-        # matplotlib figure renders to memory, which the sandbox allows.
-        _plt.matplotlib().savefig(_buf, format='png')
-        _buf.seek(0)
-        base64.b64encode(_buf.read()).decode('ascii')
-        """
-        )
-    )
-    result = await _evaluate_structured(session, code)
-    return {"image_base64": result, "format": "png"}
-
-
-@mcp.tool(
-    description="Vector calculus operations: gradient, divergence, curl, laplacian"
-)
-async def vector_calculus_operation(
-    operation: Annotated[
-        str, Field(description="One of: gradient, divergence, curl, laplacian")
-    ],
-    expression: Annotated[
-        str | list[str],
-        Field(
-            description="Scalar field (string) for gradient/laplacian, "
-            "or vector field components (list) for divergence/curl"
-        ),
-    ],
-    variables: Annotated[
-        list[str] | None,
-        Field(description="Variable names (e.g. ['x', 'y', 'z'])"),
-    ] = None,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    operation = operation.strip()
-    if variables is None:
-        variables = ["x", "y", "z"]
-    # Also quoted into var('...') below, so gate them here rather than relying on
-    # whichever branch happens to call _sage_prelude.
-    variables = [_validated_identifier(v, "variables") for v in variables]
-    session = await runtime.resolve_session(ctx.session_id, session)
-    vars_str = ", ".join(f"var('{v}')" for v in variables)
-
-    if operation == "gradient":
-        if not isinstance(expression, str):
-            raise ToolError("Gradient requires a scalar expression (string)")
-        code = (
-            _sage_prelude(variables)
-            + textwrap.dedent(
-                f"""
-            _vars = [{vars_str}]
-            _f = sage_eval({_encode_literal(expression)}, locals=_locals)
-            [str(diff(_f, v)) for v in _vars]
-            """
-            )
-        )
-    elif operation == "divergence":
-        if not isinstance(expression, list):
-            raise ToolError("Divergence requires a vector field (list of component strings)")
-        if len(expression) != len(variables):
-            raise ToolError(
-                f"Vector field has {len(expression)} components "
-                f"but {len(variables)} variables"
-            )
-        code = (
-            _sage_prelude(variables)
-            + textwrap.dedent(
-                f"""
-            _vars = [{vars_str}]
-            _components = [sage_eval(c, locals=_locals) for c in {_encode_literal(expression)}]
-            str(sum(diff(_components[i], _vars[i]) for i in range(len(_vars))))
-            """
-            )
-        )
-    elif operation == "curl":
-        if not isinstance(expression, list) or len(expression) != 3:
-            raise ToolError("Curl requires exactly 3 vector field components")
-        if len(variables) != 3:
-            raise ToolError("Curl requires exactly 3 variables")
-        code = (
-            _sage_prelude(variables)
-            + textwrap.dedent(
-                f"""
-            _vars = [{vars_str}]
-            _F = [sage_eval(c, locals=_locals) for c in {_encode_literal(expression)}]
-            _curl = [
-                str(diff(_F[2], _vars[1]) - diff(_F[1], _vars[2])),
-                str(diff(_F[0], _vars[2]) - diff(_F[2], _vars[0])),
-                str(diff(_F[1], _vars[0]) - diff(_F[0], _vars[1])),
-            ]
-            _curl
-            """
-            )
-        )
-    elif operation == "laplacian":
-        if not isinstance(expression, str):
-            raise ToolError("Laplacian requires a scalar expression (string)")
-        code = (
-            _sage_prelude(variables)
-            + textwrap.dedent(
-                f"""
-            _vars = [{vars_str}]
-            _f = sage_eval({_encode_literal(expression)}, locals=_locals)
-            str(sum(diff(_f, v, 2) for v in _vars))
-            """
-            )
-        )
-    else:
-        raise ToolError(
-            f"Unknown operation '{operation}'. "
-            "Use: gradient, divergence, curl, laplacian"
-        )
-
-    result = await _evaluate_structured(session, code)
-    return {"operation": operation, "result": result}
-
-
-@mcp.tool(description="Plot an expression and return a base64-encoded PNG image")
-async def plot_expression(
-    expression: Annotated[str, Field(description="Expression to plot")],
-    variable: Annotated[str, Field(description="Plot variable")] = "x",
-    range_min: Annotated[float, Field(description="Lower bound of plot range")] = -10.0,
-    range_max: Annotated[float, Field(description="Upper bound of plot range")] = 10.0,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required for stateful execution")
-    session = await runtime.resolve_session(ctx.session_id, session)
-    code = (
-        _sage_prelude([variable])
-        + textwrap.dedent(
-            f"""
-        import base64
-        import io as _io
-        _var = var({_encode_literal(variable)})
-        _expr = sage_eval({_encode_literal(expression)}, locals=_locals)
-        _plt = plot(_expr, (_var, {range_min}, {range_max}))
-        _buf = _io.BytesIO()
-        # Graphics.save() needs a filesystem path and rejects a BytesIO with
-        # "expected str, bytes or os.PathLike object". Going through the
-        # matplotlib figure renders to memory, which the sandbox allows.
-        _plt.matplotlib().savefig(_buf, format='png')
-        _buf.seek(0)
-        base64.b64encode(_buf.read()).decode('ascii')
-        """
-        )
-    )
-    result = await _evaluate_structured(session, code)
-    return {"image_base64": result, "format": "png"}
-
-
-# ---------------------------------------------------------------------------
-# Phase 4 — Niche domain tools
-# ---------------------------------------------------------------------------
-
-
-@mcp.tool(
-    description="Graph theory: create named graphs and compute properties "
-    "(chromatic_number, is_connected, diameter, etc.)"
-)
-async def graph_operation(
-    graph: Annotated[
-        str,
-        Field(
-            description="Graph constructor: a named graph like 'PetersenGraph' "
-            "or an adjacency dict like '{0:[1,2], 1:[0,2], 2:[0,1]}'"
-        ),
-    ],
-    operation: Annotated[
-        str,
-        Field(
-            description="One of: chromatic_number, is_connected, is_planar, "
-            "diameter, order, size, degree_sequence, adjacency_matrix, "
-            "shortest_path (requires source and target)"
-        ),
-    ],
-    source: Annotated[int | None, Field(description="Source vertex")] = None,
-    target: Annotated[int | None, Field(description="Target vertex")] = None,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required")
-    operation = operation.strip()
-    session = await runtime.resolve_session(ctx.session_id, session)
-    # A named graph is an identifier, optionally already called with arguments.
-    # Matching on a "Graph" suffix missed every parameterised constructor:
-    # "CompleteGraph(4)" ends in ")", so it fell through to Graph(CompleteGraph(4))
-    # and failed with "name 'CompleteGraph' is not defined". Most named graphs
-    # take parameters, so that was the majority of the catalogue.
-    # Validated as an expression in its own right: this string is interpolated
-    # into code that runs under the trusted policy, where sage_eval is allowed.
-    graph = _validated_expression(graph)
-    named = _NAMED_GRAPH_RE.match(graph.strip())
-    if named:
-        call = named.group("call") or "()"
-        graph_code = f"_G = graphs.{named.group('name')}{call}"
-    else:
-        # Anything else is a literal, such as an adjacency dict.
-        graph_code = f"_G = Graph({graph})"
-    ops = {
-        "chromatic_number": "int(_G.chromatic_number())",
-        "is_connected": "bool(_G.is_connected())",
-        "is_planar": "bool(_G.is_planar())",
-        "diameter": "int(_G.diameter())",
-        "order": "int(_G.order())",
-        "size": "int(_G.size())",
-        "degree_sequence": "sorted(_G.degree_sequence(), reverse=True)",
-        "adjacency_matrix": (
-            "[[int(x) for x in row] "
-            "for row in _G.adjacency_matrix().rows()]"
-        ),
-        "shortest_path": (
-            f"list(_G.shortest_path({source}, {target}))"
-            if source is not None and target is not None
-            else "None"
-        ),
-    }
-    if operation not in ops:
-        raise ToolError(
-            f"Unknown operation '{operation}'. "
-            f"Use: {', '.join(ops)}"
-        )
-    code = _sage_prelude() + graph_code + "\n" + ops[operation] + "\n"
-    result = await _evaluate_structured(session, code)
-    return {"operation": operation, "result": result}
-
-
-@mcp.tool(
-    description="Group theory: construct groups and query properties "
-    "(order, is_abelian, center, etc.)"
-)
-async def group_operation(
-    group: Annotated[
-        str,
-        Field(
-            description="Sage group constructor, e.g. "
-            "'SymmetricGroup(5)', 'DihedralGroup(4)', "
-            "'CyclicPermutationGroup(6)', 'AlternatingGroup(5)'"
-        ),
-    ],
-    operation: Annotated[
-        str,
-        Field(
-            description="One of: order, is_abelian, is_cyclic, "
-            "center_order, conjugacy_classes_count, exponent"
-        ),
-    ],
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required")
-    operation = operation.strip()
-    session = await runtime.resolve_session(ctx.session_id, session)
-    ops = {
-        "order": "int(_G.order())",
-        "is_abelian": "bool(_G.is_abelian())",
-        "is_cyclic": "bool(_G.is_cyclic())",
-        "center_order": "int(_G.center().order())",
-        "conjugacy_classes_count": (
-            "int(len(_G.conjugacy_classes_representatives()))"
-        ),
-        "exponent": "int(_G.exponent())",
-    }
-    if operation not in ops:
-        raise ToolError(
-            f"Unknown operation '{operation}'. "
-            f"Use: {', '.join(ops)}"
-        )
-    code = _sage_prelude() + f"_G = {_validated_expression(group)}\n" + ops[operation] + "\n"
-    result = await _evaluate_structured(session, code)
-    return {"group": group, "operation": operation, "result": result}
-
-
-@mcp.tool(
-    description=(
-        "Elliptic curves over Q: rank, torsion order, discriminant, j-invariant, "
-        "conductor and generators, from Weierstrass coefficients. Prefer this "
-        "over evaluate_sage for curve invariants."
-    )
-)
-async def elliptic_curve_operation(
-    coefficients: Annotated[
-        list[int],
-        Field(
-            description="Curve coefficients [a1,a2,a3,a4,a6] or "
-            "short Weierstrass [a,b] for y^2 = x^3 + a*x + b"
-        ),
-    ],
-    operation: Annotated[
-        str,
-        Field(
-            description="One of: rank, torsion_order, discriminant, "
-            "j_invariant, conductor, gens"
-        ),
-    ],
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required")
-    operation = operation.strip()
-    session = await runtime.resolve_session(ctx.session_id, session)
-    ops = {
-        "rank": "int(_E.rank())",
-        "torsion_order": "int(_E.torsion_order())",
-        "discriminant": "str(_E.discriminant())",
-        "j_invariant": "str(_E.j_invariant())",
-        "conductor": "int(_E.conductor())",
-        "gens": "[str(p) for p in _E.gens()]",
-    }
-    if operation not in ops:
-        raise ToolError(
-            f"Unknown operation '{operation}'. "
-            f"Use: {', '.join(ops)}"
-        )
-    code = (
-        _sage_prelude()
-        + f"_E = EllipticCurve({_encode_literal(coefficients)})\n"
-        + ops[operation]
-        + "\n"
-    )
-    result = await _evaluate_structured(session, code)
-    return {"operation": operation, "result": result}
-
-
-@mcp.tool(
-    description=(
-        "Error-correcting codes: length, dimension, minimum distance, rate and "
-        "generator matrix for Hamming and generalized Reed-Solomon codes. Prefer "
-        "this over evaluate_sage for code parameters."
-    )
-)
-async def coding_theory_operation(
-    code_type: Annotated[
-        str,
-        Field(
-            description="Code constructor, e.g. "
-            # ReedSolomonCode(GF(7),3,5) was documented here but has never been
-            # a valid constructor in current Sage; it raises AttributeError.
-            "'HammingCode(GF(2),3)', "
-            "'GeneralizedReedSolomonCode(GF(7).list()[:6],3)'"
-        ),
-    ],
-    operation: Annotated[
-        str,
-        Field(
-            description="One of: length, dimension, "
-            "minimum_distance, generator_matrix, rate"
-        ),
-    ],
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required")
-    operation = operation.strip()
-    session = await runtime.resolve_session(ctx.session_id, session)
-    ops = {
-        "length": "int(_C.length())",
-        "dimension": "int(_C.dimension())",
-        "minimum_distance": "int(_C.minimum_distance())",
-        "generator_matrix": (
-            "[[int(x) for x in row] "
-            "for row in _C.generator_matrix().rows()]"
-        ),
-        "rate": "float(_C.dimension() / _C.length())",
-    }
-    if operation not in ops:
-        raise ToolError(
-            f"Unknown operation '{operation}'. "
-            f"Use: {', '.join(ops)}"
-        )
-    code = (
-        _sage_prelude()
-        + f"_C = codes.{_validated_expression(code_type)}\n"
-        + ops[operation]
-        + "\n"
-    )
-    result = await _evaluate_structured(session, code)
-    return {"operation": operation, "result": result}
-
-
-@mcp.tool(
-    description=(
-        "Boolean polynomials over GF(2): evaluate, list variables, degree, and "
-        "zero/one tests. Prefer this over evaluate_sage for boolean algebra."
-    )
-)
-async def boolean_algebra_operation(
-    expression: Annotated[
-        str,
-        Field(description="Boolean expression (e.g. 'x*y + x*z + y*z')"),
-    ],
-    operation: Annotated[
-        str,
-        Field(
-            description="One of: evaluate, variables, degree, "
-            "is_zero, is_one, reduce"
-        ),
-    ],
-    num_variables: Annotated[
-        int,
-        Field(description="Number of boolean variables", ge=1),
-    ] = 3,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required")
-    operation = operation.strip()
-    session = await runtime.resolve_session(ctx.session_id, session)
-    var_names = ", ".join(f"'x{i}'" for i in range(num_variables))
-    # The ring generators are x0, x1, ..., but the documented example uses
-    # x, y, z. Expose both spellings so either parses, rather than failing
-    # with "name 'x' is not defined" on the tool's own documented input.
-    ring_setup = (
-        f"_R = BooleanPolynomialRing({num_variables}, [{var_names}])\n"
-        f"_R.inject_variables(verbose=False)\n"
-        "_bool_locals = {str(_g): _g for _g in _R.gens()}\n"
-        "for _alias, _gen in zip(['x', 'y', 'z', 'w', 'v', 'u'], _R.gens()):\n"
-        "    _bool_locals.setdefault(_alias, _gen)\n"
-        f"_bool_expr = _R(sage_eval({_encode_literal(expression)}, "
-        "locals=_bool_locals))\n"
-    )
-    ops = {
-        "evaluate": "str(_bool_expr)",
-        "variables": "[str(v) for v in _bool_expr.variables()]",
-        "degree": "int(_bool_expr.deg())",
-        "is_zero": "bool(_bool_expr.is_zero())",
-        "is_one": "bool(_bool_expr.is_one())",
-        "reduce": "str(_bool_expr)",
-    }
-    if operation not in ops:
-        raise ToolError(
-            f"Unknown operation '{operation}'. "
-            f"Use: {', '.join(ops)}"
-        )
-    code = _sage_prelude() + ring_setup + ops[operation] + "\n"
-    result = await _evaluate_structured(session, code)
-    return {"operation": operation, "result": result}
-
-
-@mcp.tool(
-    description="Polynomial ring operations: construct rings "
-    "and compute Groebner bases, ideals, quotients"
-)
-async def polynomial_ring_operation(
-    ring_vars: Annotated[
-        list[str],
-        Field(description="Variable names, e.g. ['a', 'b', 'c']"),
-    ],
-    polynomials: Annotated[
-        list[str],
-        Field(description="Polynomials as strings, e.g. ['a^2+b', 'b^2-1']"),
-    ],
-    operation: Annotated[
-        str,
-        Field(
-            description="One of: groebner_basis, ideal_dimension, "
-            "ideal_variety, reduce, is_groebner"
-        ),
-    ],
-    base_ring: Annotated[str, Field(description="Base ring")] = "QQ",
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required")
-    operation = operation.strip()
-    session = await runtime.resolve_session(ctx.session_id, session)
-    ring_vars = [_validated_identifier(v, "ring_vars") for v in ring_vars]
-    var_list = ", ".join(ring_vars)
-    ops = {
-        "groebner_basis": "[str(g) for g in _I.groebner_basis()]",
-        "ideal_dimension": "int(_I.dimension())",
-        "ideal_variety": "[{str(k): str(v) for k, v in pt.items()} "
-        "for pt in _I.variety()]",
-        "reduce": (
-            f"str(_I.reduce(_R({_encode_literal(polynomials[0])})))"
-            if polynomials
-            else "''"
-        ),
-        "is_groebner": "bool(_I.basis_is_groebner())",
-    }
-    if operation not in ops:
-        raise ToolError(
-            f"Unknown operation '{operation}'. "
-            f"Use: {', '.join(ops)}"
-        )
-    polys_code = ", ".join(
-        f"_R({_encode_literal(p)})" for p in polynomials
-    )
-    code = (
-        _sage_prelude(ring_vars)
-        + f"_R = PolynomialRing({_validated_expression(base_ring)}, '{var_list}')\n"
-        + "_R.inject_variables(verbose=False)\n"
-        + f"_I = _R.ideal([{polys_code}])\n"
-        + ops[operation]
-        + "\n"
-    )
-    result = await _evaluate_structured(session, code)
-    return {"operation": operation, "result": result}
-
-
-@mcp.tool(
-    description=(
-        "Computational geometry on point sets: euclidean distance, polygon area, "
-        "polytope volume, convex hull vertices and convexity tests. Prefer this "
-        "over evaluate_sage for these."
-    )
-)
-async def geometry_operation(
-    operation: Annotated[
-        str,
-        Field(
-            description="One of: distance, polygon_area, "
-            "polytope_volume, convex_hull_vertices, is_convex"
-        ),
-    ],
-    points: Annotated[
-        list[list[float]],
-        Field(description="List of points as coordinate lists"),
-    ],
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> dict:
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required")
-    operation = operation.strip()
-    if not points:
-        raise ToolError("'points' must contain at least one point")
-    # distance previously generated the literal "None" for a single point, so
-    # the tool returned {'result': None} as though that were an answer.
-    if operation == "distance" and len(points) < 2:
-        raise ToolError(
-            f"Operation 'distance' requires two points, got {len(points)}"
-        )
-    session = await runtime.resolve_session(ctx.session_id, session)
-    pts = _encode_literal(points)
-    ops = {
-        "distance": (
-            # "**", not "^": this expression is executed as Python, where "^"
-            # is XOR. (0-3)^2 evaluates to -1, and the sum then goes negative,
-            # so sqrt() returns a complex number and float() fails.
-            f"float(sqrt(sum((a-b)**2 for a, b in "
-            f"zip({_encode_literal(points[0])}, "
-            f"{_encode_literal(points[1])}))))"
-            if len(points) >= 2
-            else "None"
-        ),
-        "polygon_area": (
-            f"float(Polyhedron(vertices={pts}).volume())"
-        ),
-        "polytope_volume": (
-            f"float(Polyhedron(vertices={pts}).volume())"
-        ),
-        "convex_hull_vertices": (
-            f"[list(v) for v in "
-            f"Polyhedron(vertices={pts}).vertices_list()]"
-        ),
-        "is_convex": (
-            f"bool(Polyhedron(vertices={pts}).is_compact())"
-        ),
-    }
-    if operation not in ops:
-        raise ToolError(
-            f"Unknown operation '{operation}'. "
-            f"Use: {', '.join(ops)}"
-        )
-    code = _sage_prelude() + ops[operation] + "\n"
-    result = await _evaluate_structured(session, code)
-    return {"operation": operation, "result": result}
-
-
-@mcp.tool(
-    description="Execute SageMath code and stream intermediate print() output "
-    "line by line. Final result is returned as usual."
-)
-async def evaluate_sage_streaming(
-    code: Annotated[str, Field(description="SageMath code to execute")],
-    timeout_seconds: Annotated[
-        float | None,
-        Field(description="Override timeout in seconds", gt=0.0),
-    ] = None,
-    session: Annotated[str, Field(description=_SESSION_ARG_DESC)] = DEFAULT_SESSION_NAME,
-    ctx: Context | None = None,
-) -> EvaluateResult:
-    """Like evaluate_sage but emits each stdout line as a progress event."""
-    if ctx is None or ctx.session_id is None:
-        raise ToolError("MCP context with session_id is required")
-    sage_session = await runtime.resolve_session(ctx.session_id, session)
-
-    # Forward each line the moment the worker produces it. This used to await
-    # the whole evaluation and only then split the accumulated stdout, so a
-    # caller saw nothing until the computation had already finished -- which is
-    # the opposite of what the tool promises.
-    emitted = 0
-
-    async def _forward(line: str) -> None:
-        nonlocal emitted
-        emitted += 1
-        if ctx is not None:
-            await ctx.report_progress(float(emitted), None, line)
-
-    worker_result = await sage_session.evaluate(
-        code,
-        want_latex=False,
-        capture_stdout=True,
-        timeout_seconds=timeout_seconds,
-        on_stdout=_forward,
-    )
-    monitoring.record_success(worker_result.elapsed_ms)
-    return EvaluateResult(
-        result_type=worker_result.result_type,
-        result=worker_result.result,
-        latex=None,
-        stdout=_truncate_stdout(worker_result.stdout),
-        elapsed_ms=worker_result.elapsed_ms,
-    )
 
 
 def _register_health_route() -> None:
